+++ teamscale-client/src/main/java/com/teamscale/client/PrioritizableTest.java
@@ -1,5 +1,7 @@
 package com.teamscale.client;
 
+import com.squareup.moshi.Json;
+
 /**
  * {@link TestDetails} with information about their partition as well as
  * tracking data used during prioritization of tests. Two instances are
@@ -12,6 +14,18 @@ public class PrioritizableTest {
 
 	/** The reason the test has been selected. */
 	public String selectionReason;
+
+	/** Partition of the test. */
+	public String partition;
+
+	/** Duration in ms. May be null if not set. */
+	public Long durationInMs;
+
+	/**
+	 * The score determined by the TIA algorithm.
+	 */
+	@Json(name = "currentScore")
+	public double score;
 
 	public PrioritizableTest(String uniformPath) {
 		this.uniformPath = uniformPath;